+++ src/ObjectMap.js
@@ -58,6 +58,17 @@ const ldmove = (arr, fromIndex, toIndex) => {
   }
 };
 
+/* --------------------- insert --------------------- */
+const ldInsert = (arr, index, value) => {
+  if (Array.isArray(arr) && index <= arr.length) {
+    arr.splice(index, 0, value);
+  } else {
+    console.warn(
+      `Attempted to insert at index ${index}, but the operation is not valid.`
+    );
+  }
+};
+
 /* --------------------- has --------------------- */
 
 // foo -->
@@ -311,15 +322,31 @@ export class ObjectMap {
   }
 
   static move(object, path, fromIndex, toIndex) {
-    console.log('Moving:', path, fromIndex, 'to', toIndex);
+    debug(`Moving: ${path}, from index ${fromIndex} to ${toIndex}`);
     // Get the array at the specified path
     const arr = ldget(object, path);
-    console.log('Array before move', JSON.stringify(arr));
+    debug(`Array before move: ${JSON.stringify(arr)}`);
 
     // Use ldmove to perform the operation
     ldmove(arr, fromIndex, toIndex);
 
-    console.log('Array after move', JSON.stringify(arr));
+    debug(`Array after move: ${JSON.stringify(arr)}`);
+  }
+
+  static insert(object, path, index, value) {
+    debug(`Inserting: ${value} at ${path}, index ${index}`);
+    // Get the array at the specified path
+    const arr = ldget(object, path);
+    debug(`Array before insert: ${JSON.stringify(arr)}`);
+
+    // Use ldInsert to perform the operation
+    if (Array.isArray(arr)) {
+      ldInsert(arr, index, value);
+    } else {
+      debug(`Path ${path} does not point to an array.`);
+    }
+
+    debug(`Array after insert: ${JSON.stringify(arr)}`);
   }
 }