+++ vitedocs/Pages/ApiReference/Arrays/ArrrayFieldInsert/ArrayFieldInsert.jsx
@@ -5,16 +5,18 @@ import Code from '../../../../YourComponents/Code';
 import { SideBySide } from '../../../../SideBySide';
 import { Info } from '../../../../Info';
 
-export default function ArrayFieldComponent({ showItem = true }) {
+export default function ArrayFieldInsert({ showItem = true }) {
   useEffect(() => {
     window.scrollTo(0, 0);
   }, []);
   return (
     <>
       <h1>
-        <code>ArrayField Swap</code>
+        <code>ArrayField Insert</code>
       </h1>
-      <Info>Sometimes you need to swap valuse around in an array.</Info>
+      <Info>
+        Sometimes you need to insert valuues in a specific posion in an array.
+      </Info>
       <SideBySide
         leftHeader={<h3>Example:</h3>}
         rightHeader={<h3>Code:</h3>}